fixtodo: Move secret to security/pkg/k8s/.
