+++ security/pkg/k8s/controller/workloadsecret.go
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// TODO(myidpt): Move secret to security/pkg/k8s/.
-
 package controller
 
 import (
