+++ security/tests/integration/certificateRotationTest/certificate_rotation_test.go
@@ -22,7 +22,7 @@ import (
 	"time"
 
 	"istio.io/istio/pkg/log"
-	"istio.io/istio/security/pkg/pki/ca/controller"
+	"istio.io/istio/security/pkg/k8s/controller"
 	"istio.io/istio/security/tests/integration"
 	"istio.io/istio/tests/integration/framework"
 )
